# docs/OPENAPI_SWAGGER.yaml
openapi: 3.0.3
info:
  title: PayNest API
  version: "1.0.0"
  description: >
    Swagger-style OpenAPI document for PayNest wallet, account, pricing,
    payment, tag, FX, transaction history, and receipt APIs.
servers:
  - url: http://localhost:8080
    description: Local server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/pay/U2U:
    post:
      tags: [Payments]
      summary: User-to-user wallet transfer
      requestBody:
        $ref: "#/components/requestBodies/CommonPayment"
      responses:
        "200":
          description: U2U transfer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/U2UPaymentResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/pay/calculatePricing:
    post:
      tags: [Payments, Pricing]
      summary: Calculate pricing for a payment request
      requestBody:
        $ref: "#/components/requestBodies/CommonPayment"
      responses:
        "200":
          description: Pricing computation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      pricing:
                        $ref: "#/components/schemas/PricingComputationResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/pay/INTRAWALLET:
    post:
      tags: [Payments]
      summary: Transfer between wallets of same account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntraWalletTransferRequest"
      responses:
        "200":
          description: Intra-wallet transfer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntraWalletTransferResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/pay/MERCHANTPAY:
    post:
      tags: [Payments]
      summary: Merchant payment
      requestBody:
        $ref: "#/components/requestBodies/CommonPayment"
      responses:
        "200":
          description: Merchant payment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasePaymentResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/pay/CASHIN:
    post:
      tags: [Payments]
      summary: Cash-in transaction
      requestBody:
        $ref: "#/components/requestBodies/CommonPayment"
      responses:
        "200":
          description: Cash-in result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasePaymentResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/pay/CASHOUT:
    post:
      tags: [Payments]
      summary: Cash-out transaction
      requestBody:
        $ref: "#/components/requestBodies/CommonPayment"
      responses:
        "200":
          description: Cash-out result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasePaymentResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/pay/BILLPAY:
    post:
      tags: [Payments]
      summary: Bill payment
      requestBody:
        $ref: "#/components/requestBodies/CommonPayment"
      responses:
        "200":
          description: Bill payment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillPayPaymentResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/pay/o2c/initiate:
    post:
      tags: [Payments]
      summary: Initiate operator-to-channel transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/O2CPaymentRequest"
      responses:
        "200":
          description: O2C initiation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasePaymentResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/pay/o2c/status:
    post:
      tags: [Payments]
      summary: Update O2C status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalRequest"
      responses:
        "200":
          description: O2C status update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasePaymentResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/pay/stockInitiate:
    post:
      tags: [Payments]
      summary: Initiate stock creation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockInitiateRequest"
      responses:
        "200":
          description: Stock initiation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasePaymentResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/pay/stockStatusUpdate:
    post:
      tags: [Payments]
      summary: Update stock creation status
      requestBody:
        $ref: "#/components/requestBodies/StockApproval"
      responses:
        "200":
          description: Stock status update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasePaymentResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/pay/stockReimbursementInitiate:
    post:
      tags: [Payments]
      summary: Initiate stock reimbursement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockReimbursementInitiateRequest"
      responses:
        "200":
          description: Stock reimbursement initiation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasePaymentResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/pay/stockReimbursementStatusUpdate:
    post:
      tags: [Payments]
      summary: Update stock reimbursement status
      requestBody:
        $ref: "#/components/requestBodies/StockApproval"
      responses:
        "200":
          description: Stock reimbursement status update result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasePaymentResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
...
  requestBodies:
...
    CommonPayment:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CommonPaymentRequest"
    StockApproval:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StockApprovalRequest"
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          example: SUCCESS
        code:
          type: string
          nullable: true
        message:
          type: string
      additionalProperties: true
    ApiErrorResponse:
      type: object
      properties:
        responseStatus:
          $ref: "#/components/schemas/TransactionStatus"
        code:
          type: string
          example: INVALID_TOKEN
        message:
          type: string
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
        success:
          type: boolean
          example: false
        errorCode:
          type: string
        errorMessage:
          type: string
    IdentifierType:
      type: string
      enum: [ACCOUNT_ID, MOBILE, MSISDN, LOGINID]
    AuthType:
      type: string
      enum: [PIN, PASSWORD, OTP]
    RequestGateway:
      type: string
      enum: [MOBILE, WEB]
    InitiatedBy:
      type: string
      enum: [DEBITOR, CREDITOR]
    TransactionStatus:
      type: string
      enum: [SUCCESS, FAILURE, PENDING]
    BillPaymentStatus:
      type: string
      enum: [PENDING, SUCCESS, FAILED]
...
    PaymentAccountType:
      type: string
      enum: [SUBSCRIBER, MERCHANT, AGENT, BILLER, BUSINESS]
...
    WalletType:
      type: string
      description: Enum-backed values are MAIN, BONUS, SALARY; DB may also contain system wallet types such as BANK, SC, COMMDIS, TAX.
      enum: [MAIN, BONUS, SALARY, BANK, SC, COMMDIS, TAX]
...
    PricingRuleType:
      type: string
      enum: [SERVICE_CHARGE, COMMISSION, DISCOUNT, CASHBACK]
    PricingType:
      type: string
      description: Current service validator accepts STATIC and CAMPAIGN. ALLTAGS is used as a tag key, not pricingType.
      enum: [STATIC, CAMPAIGN]
...
    Identifier:
      type: object
      required: [type, value]
      properties:
        type:
          $ref: "#/components/schemas/IdentifierType"
        value:
          type: string
    Authentication:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/AuthType"
        value:
          type: string
    Party:
      type: object
      required: [accountType, identifier, walletType]
      properties:
        accountType:
          $ref: "#/components/schemas/PaymentAccountType"
        identifier:
          $ref: "#/components/schemas/Identifier"
        walletType:
          $ref: "#/components/schemas/WalletType"
        authentication:
          $ref: "#/components/schemas/Authentication"
    TransactionInfo:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: decimal
          example: 10.00
        currency:
          type: string
          example: USD
        initiatedTransactionId:
          type: string
    CommonPaymentRequest:
      type: object
      required: [operationType, debitor, creditor, transaction]
      properties:
        operationType:
          type: string
          enum: [U2U, CASHIN, CASHOUT, MERCHANTPAY, BILLPAY]
        requestGateway:
          $ref: "#/components/schemas/RequestGateway"
        preferredLang:
          type: string
        initiatedBy:
          $ref: "#/components/schemas/InitiatedBy"
        paymentReference:
          type: string
          maxLength: 100
        comments:
          type: string
          maxLength: 300
        debitor:
          $ref: "#/components/schemas/Party"
        creditor:
          $ref: "#/components/schemas/Party"
        transaction:
          $ref: "#/components/schemas/TransactionInfo"
        metadata:
          type: object
          additionalProperties: true
        additionalInfo:
          type: object
          additionalProperties: true
    BasePaymentResponse:
      type: object
      properties:
        responseStatus:
          $ref: "#/components/schemas/TransactionStatus"
        operationType:
          type: string
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
        transactionId:
          type: string
        amount:
          type: number
          format: decimal
        currency:
          type: string
    U2UPaymentResponse:
      allOf:
        - $ref: "#/components/schemas/BasePaymentResponse"
        - type: object
          properties:
            totalAmount:
              type: number
              format: decimal
            serviceCharge:
              $ref: "#/components/schemas/ServiceChargeDetails"
            discount:
              $ref: "#/components/schemas/PricingAdjustmentDetails"
            cashback:
              $ref: "#/components/schemas/PricingAdjustmentDetails"
...
    ServiceChargeDetails:
      type: object
      properties:
        amount:
          type: number
          format: decimal
        payer:
          type: string
        currency:
          type: string
    PricingAdjustmentDetails:
      type: object
      properties:
        amount:
          type: number
          format: decimal
        party:
          type: string
        payBy:
          type: string
        currency:
          type: string
    BillPayPaymentResponse:
      allOf:
        - $ref: "#/components/schemas/BasePaymentResponse"
        - type: object
          properties:
            billStatus:
              $ref: "#/components/schemas/BillPaymentStatus"
    IntraWalletTransferRequest:
      type: object
      required: [party, sourceWalletType, targetWalletType, amount, sourceCurrency, targetCurrency]
      properties:
        requestGateway:
          $ref: "#/components/schemas/RequestGateway"
        preferredLang:
          type: string
        paymentReference:
          type: string
        comments:
          type: string
        party:
          $ref: "#/components/schemas/Party"
        sourceWalletType:
          $ref: "#/components/schemas/WalletType"
        targetWalletType:
          $ref: "#/components/schemas/WalletType"
        amount:
          type: number
          format: decimal
        sourceCurrency:
          type: string
        targetCurrency:
          type: string
        metadata:
          type: object
          additionalProperties: true
        additionalInfo:
          type: object
          additionalProperties: true
    IntraWalletTransferResponse:
      type: object
      properties:
        responseStatus:
          $ref: "#/components/schemas/TransactionStatus"
        operationType:
          type: string
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
        transactionId:
          type: string
        sourceAmount:
          type: number
        sourceWalletType:
          type: string
        sourceCurrency:
          type: string
        targetAmount:
          type: number
        targetWalletType:
          type: string
        targetCurrency:
          type: string
        exchangeRate:
          type: number
        bonusToMainPercentage:
          type: number
    O2CPaymentRequest:
      type: object
      required: [operationType, channel, transaction]
      properties:
        operationType:
          type: string
          example: O2C
        requestGateway:
          $ref: "#/components/schemas/RequestGateway"
        preferredLang:
          type: string
        paymentReference:
          type: string
        comments:
          type: string
        channel:
          $ref: "#/components/schemas/Party"
        transaction:
          $ref: "#/components/schemas/TransactionInfo"
        metadata:
          type: object
          additionalProperties: true
        additionalInfo:
          type: object
          additionalProperties: true
    ApprovalRequest:
      type: object
      required: [transactionId, status]
      properties:
        transactionId:
          type: string
        status:
          type: string
        comments:
          type: string
    StockInitiateRequest:
      type: object
      required: [operationType, transaction]
      properties:
        operationType:
          type: string
        paymentReference:
          type: string
        comments:
          type: string
        transaction:
          $ref: "#/components/schemas/TransactionInfo"
        metadata:
          type: object
          additionalProperties: true
        additionalInfo:
          type: object
          additionalProperties: true
    StockApprovalRequest:
      type: object
      required: [transactionId, status]
      properties:
        transactionId:
          type: string
        status:
          type: string
        errorCode:
          type: string
        comments:
          type: string
    StockReimbursementInitiateRequest:
      type: object
      required: [debitor, transactor, transaction]
      properties:
        paymentReference:
          type: string
        comments:
          type: string
        debitor:
          $ref: "#/components/schemas/Party"
        transactor:
          $ref: "#/components/schemas/Party"
        transaction:
          $ref: "#/components/schemas/TransactionInfo"
        metadata:
          type: object
          additionalProperties: true
        additionalInfo:
          type: object
          additionalProperties: true
...
    PricingComputationResponse:
      type: object
      properties:
        senderTagKey:
          type: string
        receiverTagKey:
          type: string
        serviceChargeAffectedParty:
          type: string
        commissionAffectedParty:
          type: string
        discountAffectedParty:
          type: string
        cashbackAffectedParty:
          type: string
        serviceChargeAmount:
          type: number
        commissionAmount:
          type: number
        discountAmount:
          type: number
        cashbackAmount:
          type: number
        cashbackPayBy:
          type: string
        serviceChargeRules:
          type: array
          items:
            $ref: "#/components/schemas/PricingRuleDetails"
        commissionRules:
          type: array
          items:
            $ref: "#/components/schemas/PricingRuleDetails"
        discountRules:
          type: array
          items:
            $ref: "#/components/schemas/PricingRuleDetails"
        cashbackRules:
          type: array
          items:
            $ref: "#/components/schemas/PricingRuleDetails"
    PricingRuleDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        pricingName:
          type: string
        serviceCode:
          type: string
        ruleType:
          $ref: "#/components/schemas/PricingRuleType"
        pricingType:
          $ref: "#/components/schemas/PricingType"
        payer:
          type: string
        payBy:
          type: string
        senderTagKey:
          type: string
        receiverTagKey:
          type: string
        currency:
          type: string
        pricingConfig:
          type: object
          additionalProperties: true
        calculatedAmount:
          type: number
...